--- client/components/addPalette.1.js
@@ -1,18 +1,0 @@
-import React from 'react'
-
-function ContainedButton(props) {
-  return (
-    <div>
-      <div
-        className="paletteBtn"
-        onClick={() => {
-          props.addColor()
-        }}
-      >
-        add
-      </div>
-    </div>
-  )
-}
-
-export default ContainedButton
