+++ client/components/colorPicker.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import InputColor from 'react-input-color'
-import {ColorSelect} from './selectPalette'
+import {ColorSelect} from './addColor.js'
 
 function ColorPicker(props) {
   const [color, setColor] = React.useState({})
